+++ src/components/Exercise/ExerciseTypeContainer.tsx
@@ -23,6 +23,7 @@ const ExerciseTypeContainer: FunctionComponent<IExerciseTypeContainerProps> = ({
     const fetchExerciseData = async () => {
       try {
         const res = await getExercise(exerciseUid);
+        console.log('EXEC::', res);
         setCurrentExerciseData(res);
       } catch (e) {
         console.error(e);
